+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## 2017-07-08
 
-## 2017-17-06
+- an alternative key card_section_types.code added, constants CardSectionType::ID_INFORMATION and CardSectionType::ID_VARIANTS removed
+
+## 2017-07-06
 
 - StaticPage renamed to Page
 - an alternative key added to pages: code